# onos/sina/target/swagger.json
{"swagger":"2.0","info":{"version":"1.0.0","title":"SINA REST API","description":"APIs for SINA application."},"basePath":"/onos/sina","produces":["application/json"],"consumes":["application/json"],"tags":[{"name":"updateInfo","description":"Sina Update Information of other controller API"},{"name":"localTopology","description":"Sina Local Topology API"}],"paths":{"/updateInfo/updateData":{"post":{"tags":["updateInfo"],"summary":"Update information about other controller","description":"","parameters":[{"name":"data","in":"body","description":"String data to update","required":true}],"consumes":["application/json"],"produces":["application/json"],"responses":{"200":{"description":"successful operation"},"default":{"description":"Unexpected error"}}}},"/localTopology/getPorts":{"get":{"tags":["localTopology"],"summary":"Get list of Ports on local topology","description":"","parameters":[],"produces":["application/json"],"responses":{"200":{"description":"successful operation"},"default":{"description":"Unexpected error"}}}},"/localTopology/getLinks":{"get":{"tags":["localTopology"],"summary":"Get list of Links on local topology","description":"","parameters":[],"consumes":["application/json"],"produces":["application/json"],"responses":{"200":{"description":"successful operation"},"default":{"description":"Unexpected error"}}}},"/localTopology/exampleAPI":{"get":{"tags":["localTopology"],"summary":"Test Sina Api","description":"","parameters":[],"produces":["application/json"],"responses":{"200":{"description":"successful operation"},"default":{"description":"Unexpected error"}}}},"/localTopology/getHosts":{"get":{"tags":["localTopology"],"summary":"Get list of Hosts on local topology","description":"","parameters":[],"produces":["application/json"],"responses":{"200":{"description":"successful operation"},"default":{"description":"Unexpected error"}}}},"/localTopology/getDevices":{"get":{"tags":["localTopology"],"summary":"Get list of Devices on local topology","description":"","parameters":[],"produces":["application/json"],"responses":{"200":{"description":"successful operation"},"default":{"description":"Unexpected error"}}}}},"definitions":{}}